Update content.md
I suggest removing "takes our" and moving "current branch" to the end of the sentence to avoid passive voice. 

I suggest changing "checks to see" to "verifies". The word "see" it is not an accessible term. 

I suggest updating "weather or not" to "if" to match the guiding prinicpal: Never use a long word where a short one will do.