Update content.md
I suggest updating the header because the content discusses how to use the git commands instead of calling out the differences. 

I am not sure if the user knows that this section of content is related to Git, so I would also consider adding "Git commands" to the end for context and clarity.